server/requirements.txt
django == 4.2.1
djangorestframework==3.15.2
django-crispy-forms==2.3
django-jazzmin==0.0.0
django-taggit==6.0.0
mongoengine==0.24.1